requirements-polymarket.txt
# Python dependencies for Polymarket real trading
py-clob-client>=0.9.0
web3>=6.0.0
python-dotenv>=1.0.0
